+++ shard.go
@@ -45,6 +45,13 @@ func newShardGroup() *ShardGroup { return &ShardGroup{} }
 // Duration returns the duration between the shard group's start and end time.
 func (g *ShardGroup) Duration() time.Duration { return g.EndTime.Sub(g.StartTime) }
 
+// Contains return whether the shard group contains data for the time between min and max
+func (g *ShardGroup) Contains(min, max time.Time) bool {
+	return timeBetweenInclusive(g.StartTime, min, max) ||
+		timeBetweenInclusive(g.EndTime, min, max) ||
+		(g.StartTime.Before(min) && g.EndTime.After(max))
+}
+
 // dropSeries will delete all data with the seriesID
 func (g *ShardGroup) dropSeries(seriesID uint32) error {
 	for _, s := range g.Shards {
@@ -161,7 +168,11 @@ func (s *Shard) dropSeries(seriesID uint32) error {
 		return nil
 	}
 	return s.store.Update(func(tx *bolt.Tx) error {
-		return tx.DeleteBucket(u32tob(seriesID))
+		err := tx.DeleteBucket(u32tob(seriesID))
+		if err != bolt.ErrBucketNotFound {
+			return err
+		}
+		return nil
 	})
 }